Installation Cancel Logic
When canceling a mid-run cluster installation, set the cluster and host
states to error. From this state user will be able to reset installation.

In this commit the route logic was added along with the cluster and host
methods to set their state to error if the installation in progress.